Java updates protected for circle CI

+++ javaUpdatesOne.sh
@@ -52,7 +52,7 @@ if [[ -n $latestJavaLTS ]]; then
 
   #  Circle CI File
   f=".circleci/config.yml"
-  if [ -f $f ]; then
+  if [ -f $f ] && [ -n "$targetImageUrl" ]; then
     targetImageUrl=$(curl -s curl -s "https://api.adoptium.net/v3/assets/latest/$latestJavaLTS/hotspot" | jq -r '[.[] | select(.binary.image_type == "jdk" and .binary.os == "linux" and .binary.architecture == "x64")][0].binary.package.link')
     targetImageUrl=${targetImageUrl//\//\\/}
     sed -E 's/-\s*image:\s*gradle.*/- image: '"$targetGradleImage"'/g' "$f" > "$f""01"
